refactor: move notification funtions to <FileLoader />
