+++ src/ila26/fr.json
@@ -103,7 +103,7 @@
   "communityCreatedModal.skip": "Passer pour l'instant",
   "post": "Publier",
   "postAs": "Publier en tant que",
-  "upload.attachmentLimit": "Vous avez atteint la limite maximale de {maxFiles}",
+  "upload.attachmentLimit": "Vous avez atteint le nombre maximum de pièces jointes.",
   "upload.fileSizeLimit": "L'un des fichiers sélectionnés dépasse 1GB. Veuillez sélectionner un nouveau fichier.",
   "upload.invalidFileType": "Un ou plusieurs des fichiers sélectionnés sont du mauvais format.",
   "user.editProfile": "Modifier le profil",
